third_party/capnproto: add getFd method to fd streams
This is useful if you're receiving a generic stream and would like to
check some non-stream behavior about the file.  Of course, downcasting
can fail and code that is coupled to one particular implementation
should rethink its design.  However, it is quite handy to be able to
make syscalls on that file descriptor if available.

I've added this as prework for #26; I couldn't come up with a cleaner
design without writing my own stream class.  I'll try to send this
upstream, but I don't want to have to wait for it.