exec: refactor Applier to Apply function
Conceptually, Apply is an operation, not a stateful object.  This new
API makes it easier to hold it wrong.  Further, this enabled me to
create a new internal type: job.  jobs are created with only the state
that is necessary for a single resource, so it is much easier to
identify unnecessary shared state.